+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## [1.0.1] - 2023-03-14
+## 1.0.2 - 2023-03-17
+
+### Added
+
+- PersistentMonoBehaviourSingleton, persistent singleton between scene changes.
+
+### Fixed
+
+- MonoBehaviourSingleton internal reference released when object destroyed [#1](https://github.com/FronkonGames/GameWork-Foundation/issues/1) (_Thanks to @TJHeuvel_).
+
+## 1.0.1 - 2023-03-14
 
 ### Added
 
@@ -17,6 +27,6 @@
 - MathConstants.GoldenRatio syntax error.
 - Missing comments.
 
-## [1.0.0] - 2022-09-06
+## 1.0.0 - 2022-09-06
 
 _Initial release._